requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
Flask==2.0.2
gunicorn==20.1.0
idna==2.8
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2==2.9.3
pyTelegramBotAPI==4.4.0
python-telegram-bot==13.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.22.0
six==1.16.0
telebot==0.0.4
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.25.11
wayforpay==1.0.1
wayforpay-python==1.0.1
Werkzeug==2.0.2
